Created map-overlays.pipe as parent for others
